requirements.txt
fastapi
uvicorn
python-dotenv
openai
dspy
requests